Access control

+++ src/features/ticket/actions/delete-ticket.ts
@@ -6,12 +6,21 @@ import { setCookieByKey } from '@/actions/cookies';
 import {
   ActionState,
   fromErrorToActionState,
+  toActionState,
 } from '@/components/form/utils/to-action-state';
+import getAuthOrRedirect from '@/features/auth/queries/get-auth-or-redirect';
+import isOwner from '@/features/auth/utils/is-owner';
 import { prisma } from '@/lib/prisma';
 import { ticketsPath } from '@/paths';
 
 export default async function deleteTicket(id: string): Promise<ActionState> {
   try {
+    const { user } = await getAuthOrRedirect();
+    const ticket = await prisma.ticket.findUnique({ where: { id } });
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState('ERROR', 'Not authorized');
+    }
+
     await prisma.ticket.delete({ where: { id } });
   } catch (error) {
     return fromErrorToActionState(error);
